requirements.txt
Django==2.0
django-jalali-date==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
jdatetime==3.6.2
Pillow==7.1.2
pytz==2020.1
